langgraph/social-media-agent/requirements.txt
python-dotenv
fastapi
uvicorn
langchain-core
langchain-openai
langchain-community
langgraph
pydantic
beautifulsoup4
pytest
pytest-asyncio
httpx
arcadepy